Put sql clause builder back
